enable ia64 support
git-svn-id: http://svn.open64.net/svnroot/open64@177 ead328b1-e903-0410-a19d-eacee4b031f6

+++ projects/pathscale-2.2.1/kpro64/common/com/ia64/config_asm.h
@@ -220,8 +220,8 @@ extern char *AS_ADDRESS_UNALIGNED;
 #define ASM_DIR_ALIGN(p, s)             fprintf(Asm_File, "\t%s %d\n", AS_ALIGN,STB_align(s));
 
 /* gas .skip fills with zeroes */
-#define ASM_DIR_ZERO(fl, l)	fprintf(fl, "\t%s %d\n", AS_SPACE, l) 
-#define ASM_DIR_SKIP(fl, l) 	fprintf(fl, "\t%s %d\n", AS_SPACE, l)
+#define ASM_DIR_ZERO(fl, l)	fprintf(fl, "\t%s %lld\n", AS_SPACE, (INT64)l) 
+#define ASM_DIR_SKIP(fl, l) 	fprintf(fl, "\t%s %lld\n", AS_SPACE, (INT64)l)
   
 #ifdef __cplusplus
 }
